requirements.txt
asgiref==3.6.0
Django==4.1.7
django-countries==7.5.1
django-crispy-forms==2.0
django-filter==22.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.7
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
